backend/requirements.txt
fastapi
uvicorn
pillow
deepface
python-multipart